styling and react app added

+++ index.js
@@ -1,9 +1,11 @@
 // implement your API here
 const express = require('express');
+const cors = require('cors');
 const server = express();
 const db = require('./data/db.js');
 //midware
 server.use(express.json());
+server.use(cors());
 
 // get items
 server.get('/api/users', (req,res) => {
@@ -12,7 +14,7 @@ server.get('/api/users', (req,res) => {
             res.status(200).json(users);
         
     }).catch(err => {
-        res.status(500).json({errMessage: "err receving users from database"});
+        res.status(500).json({errorMessage: "The users information could not be retrieved."});
     });
     
 });
@@ -23,38 +25,49 @@ server.get('/api/users/:id', (req,res) => {
     db.findById(id)
         .then(user => {
             if(!user){
-                res.status(400).json({errMessage: `no user found with the ID of:${id} `});
+                res.status(404).json({ message: "The user with the specified ID does not exist."});
             }else{
                 res.status(200).json(user);
             }
         })
         .catch(err => {
             console.log(err);
-            res.status(500).json({errMessage: "Server Error while getting user"});
+            res.status(500).json({errorMessage: "The user information could not be retrieved."});
         });
 });
 
 //post items
 server.post('/api/users', (req,res) => {
     userData = req.body;
-    db.insert(userData).then( user => {
-        
-        db.findById(user.id).then( user => res.status(200).json(user)).catch(err => console.log(err));
-    }).catch(err => {
-        res.status(500).json({errMessage: "err creating user"});
-    })
+    if(!userData.name || !userData.bio ){
+        res.status(400).json({errorMessage: "Please provide name and bio for the user."});
+    }else{
+        db.insert(userData)
+            .then( user => {
+            
+                db.findById(user.id).then( user => res.status(201).json(user)).catch(err => console.log(err));
+            }).catch(err => {
+                res.status(500).json({errorMessage: "There was an error while saving the user to the database"});
+            })
+    }
 });
 //delete items
 server.delete('/api/users/:id', (req,res) => {
     const id = req.params.id;
         db.findById(id).then(user => { 
-            db.remove(id)
-            .then(resp => {
-                res.status(200).json(user);
-            }).catch(err => {
-                res.status(500).json({errMessage: "err deleteing user"});
-            });
-        }).catch(err => {res.status(500).json({errMessage: "Errr deleting User"})});
+            if(!user){
+                res.status(404).json({message: "The user with the specified ID does not exist."})
+            }else{
+       
+                db.remove(id)
+                .then(resp => {
+                    res.status(200).json(user);
+                }).catch(err => {
+                    res.status(500).json({errorMessage: "The user could not be removed"});
+                });
+            }
+       
+        }).catch(err => {res.status(500).json({message: "The user with the specified ID does not exist."})});
 
    
     
@@ -65,13 +78,21 @@ server.delete('/api/users/:id', (req,res) => {
 server.put('/api/users/:id',(req,res) => {
     const id = req.params.id;
     const data = req.body;
-
-    db.update(id,data)
-        .then(ret => {
-            res.status(201).json(ret);
+    db.findById(id)
+        .then(user => { 
+            if(!user){
+                res.status(400).json({errMessage: "The user with the specified ID does not exist."})
+            }else{
+                db.update(id,data)
+                    .then(ret => {
+                        res.status(200).json(ret);
+                    }).catch(err => {
+                        res.status(500).json({ errorMessage: "The user information could not be modified." });
+                    });
+            }
         }).catch(err => {
-            res.status(500).json({errMessage:"error updating user"});
-        });
+            res.status(500).json({ errorMessage: "The user information could not be modified." })
+    });
 });
